rename: _azimuthal_utils -> _azimuthal_integrations
